refactor: move implementation outside of interface
